Shift to using general artifacts data structure
Artifacts should include the presentation slide DOI (always present), the hands-on code repo (often present), and any other artifacts which may relate to the tutorial (particularly playlists for videos).

Because the artifact information may be needed in multiple locations (including personalizing the hands-on instructions), we need to put the artifacts in a _data file to make the universally available.