Restructure data files a bit
Looks like we'll have multiple data files per tutorial, so let's make a directory for each instead of a file.

And then, since the filename appears in the name of the data structure from Jekyll's perspective, and we can get away without naming the data structure within the file (at least for the artifacts data structure), we can avoid adding another redundant layer to the hierarchy.