Add github-metadata plugin
Using it to identify the commit used to build the site.  An experiment -- not clear how useful this will be.

Note that the last-modified-at plugin is not giving the results we wanted.  Expected to get a different date corresponding to the date the specific file was last modified in GH.  Instead we seem to get the date of the last commit or build (haven't checked that to differentiate).  This may be due to the action we're now using to build and deploy the site.  Or perhaps a deficiency or misunderstanding of the plugin.  Haven't done those experiments either yet.