Improve time zone handling
Jekyll forces everything with a time zone into the site's preferred time zone.  That's not really appropriate for tutorials, which occur all over the world.  In the absence of a way to override Jekyll's handling of times, the solution seems to be to separate dates and times.  We need to manipulate/compare dates, but we can treat times as strings.

This resolves #3 . Ticket #19 provides some ideas for future improvements.