Fix "day-of" problem
Add a day (86400 seconds) to the date of the event, so that it won't be considered past until the day is over.

(This really needs to be based on the end date of the event, but that's a bigger change.  It may also need to do a more careful job with time zones.  The date is interpreted against the time zone defined in _config.yml, but the event may be in a different time zone.)