Changes to support new staffing data structures in event.yml
In order to support dual use of event.yml files in the new ideas-productivity.org website, we need to move the people information into each of the event.yml files.  These are the essential changes that are needed to support this.  Currently, everything supports both new and old people information.  Eventually, when we convert all of the event.yml files we could probably remove support for the old format.

So far, we've tested this with one event with the new format.  So don't consider this to be completely tested yet.