Represent the old URL differently
Our URL checker looks at anything that looks like a URL, whether it is in a markdown link structure or not.  But we don't really want to put the old URL on the exclude list in case it crops up somewhere else.  So trying to make the URL not look so much like a URL.