Remove potentially problematic script reference
polyfill.io seems to have been taken over y some Chinese company and the widely-use javascript referenced there has been found to include malicious code of late.  While the reference is commented out here, let's remove it entirely to be on the safe side.  According to what I've read on the web, it should not be needed for modern browsers anyway.